Cleaned up for a first release for my account only.
The app is now fully functional, other than setting your password. If you want to use it, create DBPassword.h and add these lines to it:

    #define DRAGON_AUTH_INFO @"userid=[NICKNAME]&passwd=[PASSWORD]"
    #define DRAGON_USERID @"uid={USER ID NUMBER]"

That file is in .gitignore; don't push it up to GitHub. With that file added, though, you can compile and run the app.

Clean up changes:
* App will refresh the status every five minutes.
* Added an app icon and a version number (now 0.0.3).
* Made the menu clear out old entries correctly on refresh.
* Added a tooltip showing the current count of moves waiting.

I'm going to run it as-is for a while and then add a prefs pane for setting the password.