Rename DBStatusItem to make it a little more obvious.
NSStatusItem is the right name for the icon in the menu bar, but this class is more concerned with the menu that appears when you click that icon, rather than the icon itself.  I also think it matches better with what someone would expect to see, and with the xib name.