Track and display whether each game has been read.
The menu icon will remain full-color as long as you haven't looked at one or more of the games listed in the menu.  Games you haven't looked at yet are marked with a bullet.

The design of the status menu code is now quite a mess (particularly, it is too tightly coupled with the DBGame class). I'm also wanting tests.